Improvements on lib level
- Moved rssi function into convert module, removed rssi module
- Moved high-level public types into dedicated module
- Added explicit `#[repr(u8)]` on register's enums
- Moved set_radio_mode and set_defaults function down in the lib module
- Added sleep state for set_radio_mode
