+++ src/lowlevel/types/address_check.rs
@@ -1,6 +1,7 @@
 /// Address check configuration.
 #[allow(non_camel_case_types)]
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
+#[repr(u8)]
 pub enum AddressCheck {
     /// No address check.
     DISABLED = 0x00,
@@ -14,6 +15,6 @@ pub enum AddressCheck {
 
 impl From<AddressCheck> for u8 {
     fn from(value: AddressCheck) -> Self {
-        value as u8
+        value as Self
     }
 }
