+++ src/lowlevel/types/po_timeout.rs
@@ -1,6 +1,7 @@
 /// Programs the number of times the six-bit ripple counter must expire after XOSC has stabilized before CHP_RDYn goes low.
 #[allow(non_camel_case_types)]
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
+#[repr(u8)]
 pub enum PoTimeout {
     /// Approx. 2.3 – 2.4 μs.
     EXPIRE_COUNT_1 = 0x00,
@@ -14,6 +15,6 @@ pub enum PoTimeout {
 
 impl From<PoTimeout> for u8 {
     fn from(value: PoTimeout) -> Self {
-        value as u8
+        value as Self
     }
 }
